--- apps/dapp/src/components/Pages/Core/DappPages/BorrowPage.tsx
@@ -1,3 +1,0 @@
-export const BorrowPage = () => {
-  return <h1>Borrow goes here</h1>;
-};
